+++ android/src/main/java/com/cyanchill/missingcore/metadataretriever/MetadataRetrieverModule.kt
@@ -3,6 +3,7 @@ package com.cyanchill.missingcore.metadataretriever
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.bridge.ReactMethod
 import com.facebook.react.bridge.ReadableArray
+import com.facebook.react.bridge.ReadableMap
 import com.facebook.react.bridge.Arguments
 import com.facebook.react.bridge.Promise
 
@@ -17,7 +18,7 @@ import androidx.media3.common.util.UnstableApi
 import androidx.media3.exoplayer.MetadataRetriever
 import java.util.concurrent.ExecutionException
 
-import com.cyanchill.missingcore.metadataretriever.models.MetadataReader
+import com.cyanchill.missingcore.metadataretriever.modules.MetadataReader
 import com.cyanchill.missingcore.metadataretriever.utils.MapUtils
 import com.cyanchill.missingcore.metadataretriever.utils.Normalization
 
@@ -26,6 +27,8 @@ import com.cyanchill.missingcore.metadataretriever.utils.Normalization
 class MetadataRetrieverModule internal constructor(reactContext: ReactApplicationContext) :
   MetadataRetrieverSpec(reactContext) {
   private val context = reactContext
+
+  private var reader = MetadataReader()
 
   @ReactMethod
   override fun getMetadata(uri: String, options: ReadableArray, promise: Promise) {
@@ -55,10 +58,10 @@ class MetadataRetrieverModule internal constructor(reactContext: ReactApplicatio
         mmrMetadata.setDataSource(Normalization.getSafeUri(uri))
       }
 
-      val formatMetadataDataMap = MetadataReader.fromFormat(formatList[0])
+      val formatMetadataDataMap = reader.fromFormat(formatList[0])
       val metadataDataMap = when (mmrMetadata) {
-        null -> MetadataReader.fromMediaMetadata(mediaMetadata, wantArtwork)
-        else -> MetadataReader.fromMediaMetadataRetriever(mmrMetadata, wantArtwork)
+        null -> reader.fromMediaMetadata(mediaMetadata, wantArtwork)
+        else -> reader.fromMediaMetadataRetriever(mmrMetadata, wantArtwork)
       }
 
       var recheckBitRate = false
@@ -156,7 +159,7 @@ class MetadataRetrieverModule internal constructor(reactContext: ReactApplicatio
       if (metadataList.isEmpty()) {
         val mmrMetadata = MediaMetadataRetriever()
         mmrMetadata.setDataSource(Normalization.getSafeUri(uri))
-        promise.resolve(MetadataReader.getBase64Image(mmrMetadata.getEmbeddedPicture()))
+        promise.resolve(reader.getBase64Image(mmrMetadata.getEmbeddedPicture()))
         mmrMetadata.release()
         return
       }
@@ -176,7 +179,7 @@ class MetadataRetrieverModule internal constructor(reactContext: ReactApplicatio
           // "Other" Picture Type
           MediaMetadata.PICTURE_TYPE_OTHER -> {
             if (backupImage == null || backupImageCode == 1) {
-              val newImg = MetadataReader.getBase64Image(mediaMetadata.artworkData)
+              val newImg = reader.getBase64Image(mediaMetadata.artworkData)
               if (newImg !== null) {
                 backupImage = newImg
                 backupImageCode = 3
@@ -186,13 +189,13 @@ class MetadataRetrieverModule internal constructor(reactContext: ReactApplicatio
           // "32x32 pixels 'file icon' (PNG only)" Picture Type
           MediaMetadata.PICTURE_TYPE_FILE_ICON -> {
             if (backupImage == null) {
-              backupImage = MetadataReader.getBase64Image(mediaMetadata.artworkData)
+              backupImage = reader.getBase64Image(mediaMetadata.artworkData)
               backupImageCode = 1
             }
           }
           // "Cover (front)" Picture Type
           MediaMetadata.PICTURE_TYPE_FRONT_COVER -> {
-            coverImage = MetadataReader.getBase64Image(mediaMetadata.artworkData)
+            coverImage = reader.getBase64Image(mediaMetadata.artworkData)
           }
         }
 
@@ -211,6 +214,15 @@ class MetadataRetrieverModule internal constructor(reactContext: ReactApplicatio
     } catch (e: Exception) {
       promise.reject("ERR_ARTWORK", e.message, e)
     }
+  }
+
+  /** Expose to the user the ability to update internal configuration options. */
+  @ReactMethod
+  override fun updateConfigs(options: ReadableMap, promise: Promise) {
+    Arguments.toBundle(options)?.let {
+      reader.updateConfigs(it)
+    }
+    promise.resolve(null)
   }
 
   //#region [Internal Helpers]