+++ android/src/oldarch/MetadataRetrieverSpec.kt
@@ -3,6 +3,7 @@ package com.cyanchill.missingcore.metadataretriever
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.bridge.ReactContextBaseJavaModule
 import com.facebook.react.bridge.ReadableArray
+import com.facebook.react.bridge.ReadableMap
 import com.facebook.react.bridge.Promise
 
 abstract class MetadataRetrieverSpec internal constructor(context: ReactApplicationContext) :
@@ -10,4 +11,6 @@ abstract class MetadataRetrieverSpec internal constructor(context: ReactApplicat
   abstract fun getMetadata(uri: String, options: ReadableArray, promise: Promise)
 
   abstract fun getArtwork(uri: String, promise: Promise)
+
+  abstract fun updateConfigs(options: ReadableMap, promise: Promise)
 }