+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 ## [Unreleased]
 
+### ❗ Breaking Changes
+
+- Remove additional fallback when getting the `year` metadata field for `MetadataRetriever` method.
+  - No longer fallbacks to using `MediaMetadataRetriever` if everything else fails.
+
+### ⚙️ Internal Changes
+
+- Reorganization of code.
+  - Now "normalize" and put all the field-value pairs inside a map, which we then send the ones we want to return.
+- Ensure we release `MetadataRetriever` & `MediaMetadataRetriever` when we're done using them.
+
 ## [1.1.0] - 2025-08-21
 
 ### ⚡ Changes